adding interceptor ee version
